+++ site/Gemfile
@@ -1,0 +1,6 @@
+source 'https://rubygems.org'
+
+gem 'd-mark', path: '..'
+gem 'nanoc', '~> 4.1'
+gem 'nanoc-git', '~> 1.0'
+gem 'nokogiri'
