+++ site/content/index.dmark
@@ -1,3 +1,7 @@
+---
+title: D★Mark
+---
+
 #h D★Mark
 
 #p %firstterm{D★Mark} is a language for marking up prose. It facilitates writing semantically meaningful text, without limiting itself to the semantics provided by HTML or Markdown. If you’re a technical writer looking for a flexible markup language, D★Mark might be a good fit.
