+++ site/content/stylesheet.css
@@ -1,0 +1,140 @@
+* {
+  margin: 0;
+  padding: 0;
+}
+
+html {
+  font-size: 15px;
+  line-height: 1.5rem;
+  padding: 4rem;
+  border-left: 1rem solid #ee5070;
+}
+
+body {
+  max-width: 720px;
+  font-family: Merriweather, serif;
+  color: #444;
+  background-color: #f8fcff;
+}
+
+h1, h2, h3 {
+  font-family: Montserrat, sans-serif;
+  line-height: 1em;
+  margin: 1.5em 0 1em 0;
+  color: #ee5070;
+}
+
+h1 {
+  font-size: 48px;
+  margin-top: 0;
+}
+
+h2 {
+  font-size: 36px;
+}
+
+h3 {
+  font-size: 24px;
+}
+
+em, .firstterm {
+  font-style: italic;
+}
+
+a:link, a:visited {
+  color: rgb(77, 103, 242);
+}
+
+a:hover {
+  color: #ee5070;
+}
+
+a:active {
+  color: #000;
+}
+
+dl, p, pre, blockquote, ul, ol, .note, .todo {
+  margin: 0 0 1.5rem 0;
+}
+
+dt {
+  font-weight: bold;
+  color: #000;
+  margin: 2rem 0 1rem 0;
+}
+
+dd {
+  margin: 0 0 1em 0;
+  padding: 0 0 0 2em;
+}
+
+ul, ol {
+  margin-left: 2rem;
+}
+
+pre {
+  padding: 1em;
+  background: rgba(77, 103, 242, 0.1);
+  color: #000;
+  white-space: pre-wrap;
+}
+
+blockquote {
+  padding: 1em;
+  background: rgba(0, 0, 0, 0.05);
+  color: #000;
+  font-style: italic;
+}
+
+blockquote :last-child {
+  margin-bottom: 0;
+}
+
+code, kbd, samp, var {
+  font-size: 14px;
+  font-family: Cousine, monospace;
+}
+
+.dmark-punctuation {
+  color: #999;
+}
+
+.dmark-elem {
+  font-weight: bold;
+}
+
+.prompt {
+  color: #999;
+}
+
+kbd {
+  font-weight: bold;
+}
+
+.filename {
+  font-style: italic;
+}
+
+.note {
+  padding: 1em;
+  border: 1px solid #000;
+}
+
+.note:before {
+  content: "NOTE: ";
+  font-weight: bold;
+  font-family: Montserrat, sans-serif;
+}
+
+.todo {
+  padding: 1em;
+  color: #d00;
+  background: rgba(255, 0, 0, 0.05);
+  border: 1px solid #f00;
+}
+
+.todo:before {
+  content: "TODO: ";
+  font-weight: bold;
+  font-family: Montserrat, sans-serif;
+}
