+++ site/lib/gems.rb
@@ -1,0 +1,1 @@
+require 'nanoc-git'
